Refactor entities to use BigDecimal and OffsetDateTime. Clean up ApprovalVerification controller error handling.

+++ bidverse/src/main/java/com/example/bidverse/Entity/Room_Seat.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -23,7 +24,7 @@ public class Room_Seat {
     @Column(name = "buyer_id", nullable = false)
     private Long buyerId;
     @Column(name = "advance_amount", nullable = false)
-    private Double advanceAmount;
+    private BigDecimal advanceAmount;
     @Column(name = "advance_status", nullable = false)
     private String advance_status = "booked";
